% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{forest_plot}
\alias{forest_plot}
\title{Creates a forest plot to visualise partial genetic correlation results}
\usage{
forest_plot(
  obj,
  save_file = F,
  file_name = NULL,
  file_width = NULL,
  conditions = NULL,
  pairs = NULL,
  order = "difference"
)
}
\arguments{
\item{obj}{an object containing the output from a call to \code{\link{partial_ldsc}()}}

\item{save_file}{Should the graphic should be saved?
\code{default=FALSE}, graphic will be displayed on the on-screen device (logical)}

\item{file_name}{The name of the file to be saved (if \code{save_file} is \code{TRUE}),
\code{default=NULL}, will use <forest_plot.jpeg>, extensions allowed are ".jpeg"/".pdf"/".tiff" (character)}

\item{file_width}{The width of the figure to be saved (if \code{save_file} is \code{TRUE}),
to help with readability if the conditions' names are too long, \code{default=NULL} (numeric)}

\item{conditions}{The names of the conditions to be included, \code{default=NULL},
will use all conditions and all corresponding pairs, a maximum of 8 conditions (character)}

\item{pairs}{The names of the pairs to be included, \code{default=NULL},
if provided, will overrule \code{conditions}, a maximum of 30 pairs (character)}

\item{order}{How to order the conditions,  possibility to use the difference (\code{"difference"}),
the unadjusted genetic correlation (\code{"unadjusted"}), or the partial genetic correlation (\code{"partial"}),
\code{default="difference"} (character)}
}
\value{
a Forest plot
}
\description{
Creates a forest plot to visualise partial genetic correlation results
}
