% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{get_conditions}
\alias{get_conditions}
\title{Get list of conditions from partial_ldsc results}
\usage{
get_conditions(x, ...)
}
\arguments{
\item{x}{an object containing the output from a call to \code{\link{partial_ldsc}()}}

\item{...}{further arguments}
}
\value{
the names of all the conditions from the analysis
}
\description{
Get list of conditions from partial_ldsc results
}
