% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{get_confounders}
\alias{get_confounders}
\title{Get list of confounders from partial_ldsc results}
\usage{
get_confounders(x, ...)
}
\arguments{
\item{x}{an object containing the output from a call to \code{\link{partial_ldsc}()}}

\item{...}{further arguments}
}
\value{
the names of all the confounders from the analysis
}
\description{
Get list of confounders from partial_ldsc results
}
