% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{get_pairs}
\alias{get_pairs}
\title{Get list of pairs from partial_ldsc results}
\usage{
get_pairs(x, ...)
}
\arguments{
\item{x}{an object containing the output from a call to \code{\link{partial_ldsc}()}}

\item{...}{further arguments}
}
\value{
the names of all the pairs from the analysis
}
\description{
Get list of pairs from partial_ldsc results
}
