% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldsc_partial.R
\name{partial_ldsc}
\alias{partial_ldsc}
\title{ldsc_partial - main function to estimate partial genetic correlations}
\usage{
partial_ldsc(
  conditions,
  confounder,
  condition.names = NULL,
  confounder.name = NULL,
  ld,
  n.blocks = 200,
  ldsc.log = NULL
)
}
\arguments{
\item{conditions}{The path to the files containing munged the GWAS summary statistics
for the conditions (character)}

\item{confounder}{The path to the file containing munged the GWAS summary statistics
for the confounder (character)}

\item{condition.names}{The names of the conditions, \code{default=NULL} (character)}

\item{confounder.name}{The name of the confounder, \code{default=NULL} (character)}

\item{ld}{The path to the folder in which the LD scores used in the analysis are located.
Expects LD scores formated as required by the original LD score regression software.  (character)}

\item{n.blocks}{The number of blocks used for block-jackknife, \code{default=200} (numeric)}

\item{ldsc.log}{The name of the log file, \code{default=NULL} (character)}
}
\description{
Estimates unadjusted and partial genetic correlations (and heritabilities, on observed scale),
and tests for the difference between the two, for all pairs of conditions.
}
\details{
\code{conditions} and \code{confounder} are required arguments.
These input files should have been pre-processed using the \code{munge()} function
from \href{https://github.com/GenomicSEM/GenomicSEM}{\code{GenomicSEM}}
}
